SDL2 should always go to the "SDL2"-subfolder.
If that breaks packages, **they** need to substitute either their "#include" (preferably) or adapt their include paths to look into this subfolder. We fix this for them by setting the include path to the subfolder (making the assumption that people wont try to include both: SDL1 and 2)

But normally when you switch between incompatible versions of a library, you'd need to adapt some stuff. We're too generous!